@@ -1,5 +1,7 @@
 import { Injectable } from "@angular/core"
-import { Http } from "@angular/http"
+
+// modulo HttpClient ja trabalha com padrao JSON
+import { HttpClient, HttpParams } from "@angular/common/http"
 
 import { Observable } from "rxjs/Observable"
 import 'rxjs/add/operator/map'
@@ -8,36 +10,32 @@ import 'rxjs/add/operator/catch'
 import { Restaurant } from "./restaurant/restaurant.model"
 
 import { MEAT_API } from "../app.api"
-import { ErrorHandler } from "../app.error-handler"
 import { MenuItem } from "app/restaurant-detail/menu-item/menu-item.model"
 
 @Injectable()
 export class RestaurantsService {
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   restaurants(search?: string): Observable<Restaurant[]> {
-    return this.http.get(`${MEAT_API}/restaurants`, { params: { q: search } })
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+    let params: HttpParams = undefined
+    if (search) {
+      // o parametro precisa ser setado no momento da instancia.. caso contrario ele cria uma copia
+      params = new HttpParams().set('q', search)
+    }
+    return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, { params: params })
   }
 
   restaurantById(id: string): Observable<Restaurant> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+    return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
   }
 
   reviewOfRestaurant(id: string): Observable<any> {
     return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
   }
 
   menuOfRestaurant(id: string): Observable<MenuItem[]> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError)
+    return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
   }
 
 }
